Refactor: Separate the concerns of display.go
